+++ device-bootloaders/src/ServiceProcessorBL2/driver/minion_configuration.c
@@ -261,12 +261,12 @@ static StaticTimer_t MM_Timer_Buffer;
 /*! \def THROTTLE_FREQUENCY_STEP
     \brief The change in frequency per step in megahertz.
 */
-#define THROTTLE_FREQUENCY_STEP 50.0
+#define THROTTLE_FREQUENCY_STEP 50U
 
 /*! \def THROTTLE_VOLTAGE_STEP_MV
     \brief The change in voltage per step in millivolts.
 */
-#define THROTTLE_VOLTAGE_STEP_MV 10.0
+#define THROTTLE_VOLTAGE_STEP_MV 10U
 
 static uint64_t gs_active_shire_mask = 0;
 static uint64_t gs_dlls_initialized = 0;